+++ server/plugins/messageStorage/types.d.ts
@@ -5,6 +5,7 @@ import {Message} from "../../models/message";
 import {Network} from "../../models/network";
 import Client from "../../client";
 import type {MessageType} from "../../../shared/types/msg";
+import {SearchQuery} from "../../../shared/types/storage";
 
 export type DeletionRequest = {
 	olderThanDays: number;
@@ -27,13 +28,6 @@ interface MessageStorage {
 
 	canProvideMessages(): boolean;
 }
-
-export type SearchQuery = {
-	searchTerm: string;
-	networkUuid: string;
-	channelName: string;
-	offset: number;
-};
 
 export type SearchResponse = SearchQuery & {
 	results: Message[];
