+++ shared/types/storage.ts
@@ -1,0 +1,6 @@
+export type SearchQuery = {
+	searchTerm: string;
+	networkUuid: string;
+	channelName: string;
+	offset: number;
+};
